+++ app/src/main/java/com/concordium/wallet/ui/MainActivity.kt
@@ -192,6 +192,11 @@ class MainActivity : BaseActivity(R.layout.activity_main, R.string.accounts_over
             finishAffinity()
             startActivity(Intent(this, MainActivity::class.java))
         }
+        walletConnectViewModel.eventsFlow.collectWhenStarted(this) {
+            if (it == WalletConnectViewModel.Event.GoBack) {
+               moveTaskToBack(true)
+            }
+        }
     }
 
     private fun initializeViews() {
@@ -305,7 +310,7 @@ class MainActivity : BaseActivity(R.layout.activity_main, R.string.accounts_over
         val walletConnectUri = intent.getStringExtra(EXTRA_WALLET_CONNECT_URI)
             ?.takeIf(String::isNotEmpty)
         if (walletConnectUri != null) {
-            walletConnectViewModel.handleWcUri(walletConnectUri)
+            walletConnectViewModel.handleUri(walletConnectUri)
         }
     }