+++ app/src/main/java/com/concordium/wallet/ui/welcome/WelcomeAccountActivationLauncher.kt
@@ -1,5 +1,0 @@
-package com.concordium.wallet.ui.welcome
-
-interface WelcomeAccountActivationLauncher {
-    fun proceedWithAccountActivation()
-}