+++ app/src/main/java/com/concordium/wallet/ui/welcome/WelcomeActivity.kt
@@ -2,37 +2,24 @@ package com.concordium.wallet.ui.welcome
 
 import android.app.Activity
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.content.ContextCompat
+import androidx.fragment.app.FragmentTransaction
+import androidx.fragment.app.commit
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.get
 import com.concordium.wallet.App
 import com.concordium.wallet.R
-import com.concordium.wallet.databinding.ActivityWelcomeBinding
-import com.concordium.wallet.extension.collectWhenStarted
-import com.concordium.wallet.extension.showSingle
 import com.concordium.wallet.ui.MainActivity
 import com.concordium.wallet.ui.auth.setup.AuthSetupPasscodeActivity
 import com.concordium.wallet.ui.base.BaseActivity
-import com.concordium.wallet.ui.more.notifications.NotificationsPermissionDialog
-import com.concordium.wallet.uicore.handleUrlClicks
 
-class WelcomeActivity :
-    BaseActivity(R.layout.activity_welcome),
-    WelcomeAccountActivationLauncher {
-
-    private val binding: ActivityWelcomeBinding by lazy {
-        ActivityWelcomeBinding.bind(findViewById(R.id.root_layout))
-    }
-
-    private val trackingPreferences = App.appCore.appTrackingPreferences
+class WelcomeActivity : BaseActivity(R.layout.activity_welcome) {
 
     private val passcodeSetupForCreateLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
             if (it.resultCode == Activity.RESULT_OK) {
-                gotoAccountOverview()
+                goToAccountOverview()
             }
         }
     private val passcodeSetupForImportLauncher =
@@ -54,32 +41,31 @@ class WelcomeActivity :
 
         initViews()
 
-        viewModel.isNotificationDialogEverShowed.collectWhenStarted(this) {
-            if (it.not()) {
-                NotificationsPermissionDialog().showSingle(
-                    supportFragmentManager,
-                    NotificationsPermissionDialog.TAG,
-                )
-                viewModel.setNotificationDialogShowed()
-            }
-        }
-
-        // Subscribe to activation bottom sheet chosen action.
         supportFragmentManager.setFragmentResultListener(
-            WelcomeActivateAccountBottomSheet.ACTION_REQUEST,
+            WelcomeGetStartedFragment.ACTION_REQUEST,
             this
         ) { _, bundle ->
-            when (WelcomeActivateAccountBottomSheet.getResult(bundle)) {
-                WelcomeActivateAccountBottomSheet.ChosenAction.CREATE ->
+            when (WelcomeGetStartedFragment.getChosenAction(bundle)) {
+                WelcomeGetStartedFragment.ChosenAction.CREATE ->
                     if (viewModel.shouldSetUpPassword) {
-                        // Set up password and create a new wallet.
+                        // Set up password and create a new seed phrase wallet.
                         passcodeSetupForCreateLauncher
-                            .launch(Intent(this, AuthSetupPasscodeActivity::class.java))
+                            .launch(
+                                Intent(this, AuthSetupPasscodeActivity::class.java)
+                                    .putExtra(
+                                        AuthSetupPasscodeActivity.SET_UP_SEED_PHRASE_WALLET_EXTRA,
+                                        true
+                                    )
+                            )
                     } else {
-                        gotoAccountOverview()
+                        // If the password was created for Import, but now they returned
+                        // and want to Create, initial setup must be marked as finished
+                        // to avoid getting back again.
+                        App.appCore.setup.finishInitialSetup()
+                        goToAccountOverview()
                     }
 
-                WelcomeActivateAccountBottomSheet.ChosenAction.IMPORT ->
+                WelcomeGetStartedFragment.ChosenAction.IMPORT ->
                     if (viewModel.shouldSetUpPassword) {
                         // Set up password and create a import an existing wallet.
                         passcodeSetupForImportLauncher
@@ -87,6 +73,20 @@ class WelcomeActivity :
                     } else {
                         goToImportWallet()
                     }
+            }
+        }
+
+        supportFragmentManager.setFragmentResultListener(
+            WelcomeCarouselFragment.GET_STARTED_REQUEST,
+            this
+        ) { _, _ ->
+            supportFragmentManager.commit {
+                replace(
+                    R.id.fragment_container,
+                    WelcomeGetStartedFragment()
+                )
+                setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
+                disallowAddToBackStack()
             }
         }
     }
@@ -97,60 +97,18 @@ class WelcomeActivity :
     }
 
     private fun initViews() {
-        binding.termsTextView.handleUrlClicks { url ->
-            when (url) {
-                "#terms" ->
-                    openTerms()
 
-                else -> {
-                    val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(url))
-                    ContextCompat.startActivity(this, browserIntent, null)
-                }
-            }
-        }
-
-        binding.termsCheckBox.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) {
-                App.appCore.tracker.welcomeTermAndConditionsCheckBoxChecked()
-            }
-            binding.getStartedButton.isEnabled = isChecked
-        }
-
-        binding.activityTrackingCheckBox.isChecked = trackingPreferences.isTrackingEnabled
-        binding.activityTrackingCheckBox.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) {
-                App.appCore.tracker.welcomeActivityTrackingCheckBoxChecked()
-            }
-            trackingPreferences.isTrackingEnabled = isChecked
-        }
-
-        binding.getStartedButton.setOnClickListener {
-            App.appCore.tracker.welcomeGetStartedClicked()
-            proceedWithAccountActivation()
-        }
-    }
-
-    private fun openTerms() {
-        startActivity(Intent(this, WelcomeTermsActivity::class.java))
     }
 
     override fun loggedOut() {
         // do nothing as we are one of the root activities
     }
 
-    override fun proceedWithAccountActivation() {
-        WelcomeActivateAccountBottomSheet()
-            .show(supportFragmentManager, WelcomeActivateAccountBottomSheet.TAG)
-    }
-
     private fun goToImportWallet() {
         startActivity(Intent(this, WelcomeRecoverWalletActivity::class.java))
     }
 
-    private fun gotoAccountOverview() {
-        // hasCompletedInitialSetup has positive value at the fresh app start,
-        // hence we can skip this screen if the user created the password.
-        App.appCore.setup.finishInitialSetup()
+    private fun goToAccountOverview() {
         finish()
         val intent = Intent(this, MainActivity::class.java)
         intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)